Introduce fumadocs-wikilink, an opt-in remark plugin for Obsidian-style wikilinks, with first-party documentation. Update pnpm-lock.yaml and package.json to include the new package, and integrate it into the docs configuration.
