+++ .changeset/silver-files-sneeze.md
@@ -1,6 +1,5 @@
 ---
-'fumadocs-core': major
-'fumadocs-mdx': major
+'fumadocs-wikilink': minor
 ---
 
-Wikilink Support
+Introduce fumadocs-wikilink — an opt-in remark plugin for Obsidian-style wikilinks. First-party docs enable it; no default behavior changes to core or mdx packages.
