+++ apps/docs/content/docs/mdx/mdx.mdx
@@ -19,6 +19,10 @@ These plugins are applied by default:
 - [Remark Structure](/docs/headless/mdx/structure) - Generate search indexes
 - Remark Exports - Exports the output generated by remark plugins above
 
+Optional remark plugins:
+
+- [Wikilink Support](/docs/headless/mdx/remark-wikilink) - Support wikilinks (opt-in)
+
 Add other remark plugins with:
 
 ```ts tab="Global Config"
@@ -97,6 +101,9 @@ export default defineConfig({
     remarkHeadingOptions: {
       // options
     },
+    // Optional: enable Wikilink support via plugin
+    // import { remarkWikilink } from 'fumadocs-wikilink'
+    // remarkPlugins: [remarkWikilink],
   },
 });
 ```
