+++ packages/mdx/src/config/build.ts
@@ -6,9 +6,6 @@ import type {
   MetaCollection,
 } from '@/config/define';
 import type { ProcessorOptions } from '@mdx-js/mdx';
-import * as fs from 'node:fs';
-import * as path from 'node:path';
-import { fumaMatter } from '@/utils/fuma-matter';
 
 export function buildConfig(config: Record<string, unknown>): LoadedConfig {
   const collections: LoadedConfig['collections'] = new Map();
@@ -56,19 +53,8 @@ export function buildConfig(config: Record<string, unknown>): LoadedConfig {
 
         if (options?.preset === 'minimal') return options;
 
-        const inferredKeyMap = buildKeyLinkMapFromDocs();
-        const withWikilink = {
-          ...options,
-          // preserve user override if provided
-          remarkWikilinkOptions:
-            options?.remarkWikilinkOptions ??
-            (Object.keys(inferredKeyMap).length > 0
-              ? { keyLinkMap: inferredKeyMap }
-              : undefined),
-        } as Parameters<typeof getDefaultMDXOptions>[0];
-
         return getDefaultMDXOptions({
-          ...withWikilink,
+          ...(options as Parameters<typeof getDefaultMDXOptions>[0]),
           _withoutBundler: mode === 'remote',
         });
       }
@@ -78,65 +64,4 @@ export function buildConfig(config: Record<string, unknown>): LoadedConfig {
       return result;
     },
   };
-}
-
-function buildKeyLinkMapFromDocs(): Record<string, string> {
-  // Scan conventional docs directories and build a key -> URL map.
-  const cwd = process.cwd();
-  const candidates = [
-    path.join(cwd, 'apps/docs/content/docs'),
-    path.join(cwd, 'content/docs'),
-  ];
-
-  const root = candidates.find((p) => fs.existsSync(p));
-  if (!root) return {};
-
-  const map: Record<string, string> = {};
-
-  function scan(dir: string, segments: string[] = []): void {
-    for (const entry of fs.readdirSync(dir, { withFileTypes: true })) {
-      if (entry.isDirectory()) {
-        // Skip route groups like (group)
-        if (entry.name.startsWith('(')) continue;
-        scan(path.join(dir, entry.name), [...segments, encodeURI(entry.name)]);
-        continue;
-      }
-
-      if (!entry.name.endsWith('.mdx')) continue;
-
-      const filePath = path.join(dir, entry.name);
-      const base = path.basename(entry.name, '.mdx');
-      const slugs =
-        base === 'index' ? [...segments] : [...segments, encodeURI(base)];
-      const url = `/docs/${slugs.join('/')}`;
-
-      try {
-        const raw = fs.readFileSync(filePath, 'utf8');
-        const data = fumaMatter(raw).data as
-          | {
-              title?: unknown;
-              aliases?: unknown;
-              slug?: unknown;
-            }
-          | undefined;
-
-        const keys: string[] = [];
-        if (typeof data?.title === 'string') keys.push(data.title);
-        if (Array.isArray(data?.aliases)) {
-          for (const a of data.aliases)
-            if (typeof a === 'string' && a) keys.push(a);
-        }
-        if (typeof data?.slug === 'string') keys.push(data.slug);
-        const last = slugs[slugs.length - 1];
-        if (last) keys.push(last.replace(/[-_]/g, ' '));
-
-        for (const key of keys) map[key] = url;
-      } catch {
-        // ignore parse/frontmatter errors and continue
-      }
-    }
-  }
-
-  scan(root);
-  return map;
 }
